+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-04-27T03:19:16.956Z")
+    updated: Optional[datetime] = parse_datetime("2026-04-27T08:21:11.069Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")