+++ packages/orq-rc/RELEASES.md
@@ -7339,3 +7339,13 @@ Based on:
 - [python v4.9.0-rc.1] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.1] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.1 - packages/orq-rc
+
+## 2026-04-27 08:24:03
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.2] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.2] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.2 - packages/orq-rc